FF-489 #comment Started cleaning up code by commenting and split out owl rdf parsing and retrieval from rest of generate_ontology; will make subsequent changes on top of this branch so lots of changes to come
